backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client for MCP calls
httpx==0.25.2

# Qdrant vector database
qdrant-client==1.8.0

# AI Guardrails
guardrails-ai==0.4.5

# Environment management
python-dotenv==1.0.0

# Structured logging
structlog==23.2.0

# Data processing and embeddings (reusing from database module)
sentence-transformers==2.2.2
datasets==2.18.0
pandas==2.1.4

# MCP client (for web search integration)
fastmcp==0.3.0

# Logging and monitoring
structlog==23.2.0

# Data validation
pydantic==2.5.0

# Async support
asyncio==3.4.3

# UUID generation (built-in, but listed for clarity)
# uuid (built-in)

# JSON handling
orjson==3.9.10
